+++ jgclark.SearchExtensions/src/searchHelpers.js
@@ -2,24 +2,20 @@
 // @flow
 //-----------------------------------------------------------------------------
 // Search Extensions helpers
+// Note: some types + funcs now in @helpers/extendedSearch.js
 // Jonathan Clark
-// Last updated 2025-02-22 for v1.5.0, @jgclark
+// Last updated 2025-03-01 for v1.5.0, @jgclark
 //-----------------------------------------------------------------------------
 
 import pluginJson from '../plugin.json'
-import {
-  getDateStrForStartofPeriodFromCalendarFilename,
-  withinDateRange,
-} from '@helpers/dateTime'
-import {
-  nowLocaleShortDateTime,
-} from '@helpers/NPdateTime'
-import { eliminateDuplicateSyncedParagraphs } from '@helpers/syncedCopies'
+import { nowLocaleShortDateTime } from '@helpers/NPdateTime'
+import { getDateStrForStartofPeriodFromCalendarFilename, withinDateRange } from '@helpers/dateTime'
 import { clo, logDebug, logError, logInfo, logTimer, logWarn, timer } from '@helpers/dev'
 import {
   displayTitle,
   type headingLevelType,
 } from '@helpers/general'
+import { differenceByPropVal, differenceByObjectEquality } from '@helpers/dataManipulation'
 import {
   getNoteByFilename, getNoteContextAsSuffix, getOrMakeNote,
   replaceSection
@@ -28,6 +24,7 @@ import { getNoteTitleFromFilename } from '@helpers/NPnote'
 import { isTermInMarkdownPath, isTermInURL } from '@helpers/paragraph'
 import { fullWordMatch, trimAndHighlightTermInLine } from '@helpers/search'
 import { sortListBy } from '@helpers/sorting'
+import { eliminateDuplicateSyncedParagraphs } from '@helpers/syncedCopies'
 import { showMessageYesNo } from '@helpers/userInput'
 
 //------------------------------------------------------------------------------
@@ -47,33 +44,34 @@ export type typedSearchTerm = {
   type: 'must' | 'may' | 'not-line' | 'not-note' | 'regex',
 }
 
-export type resultObjectTypeV3 = {
+export type resultObjectType = {
   searchTerm: typedSearchTerm,
   resultNoteAndLineArr: Array<noteAndLine>,
   resultCount: number,
 }
 
-export type resultOutputTypeV3 = {
-  searchTermsRepArr: Array<string>,
-  resultNoteAndLineArr: Array<noteAndLine>,
-  resultCount: number,
-  resultNoteCount: number,
-  fullResultCount: number,
-}
+export type resultOutputType = {
+  searchTermsRepArr: Array<string>;
+  resultNoteAndLineArr: Array<noteAndLine>;
+  resultCount: number;
+  resultNoteCount: number;
+  fullResultCount: number;
+};
 
 // Reduced set of paragraph.* fields
 export type reducedFieldSet = {
-  filename: string,
-  changedDate?: Date,
-  createdDate?: Date,
-  title: string,
-  type: ParagraphType,
-  content: string,
-  rawContent: string,
-  lineIndex: number,
+  filename: string;
+  changedDate?: Date;
+  createdDate?: Date;
+  title: string;
+  type: ParagraphType;
+  content: string;
+  rawContent: string;
+  lineIndex: number;
 }
 
 //-------------------------------------------------------------------------------
+// Constants
 
 export const OPEN_PARA_TYPES = ['open', 'scheduled', 'checklist', 'checklistScheduled']
 export const SYNCABLE_PARA_TYPES = ['open', 'scheduled', 'checklist', 'checklistScheduled']
@@ -129,6 +127,16 @@ export async function getSearchSettings(): Promise<any> {
 }
 
 //------------------------------------------------------------------------------
+// Functions
+
+/**
+ * Get string representation of multiple search terms, complete with surrounding sqaure brackets (following Google's style)
+ * @param {typedSearchTerm[]} searchTerms
+ * @returns {string}
+ */
+export function getSearchTermsRep(typedSearchTerms: Array<typedSearchTerm>): string {
+  return `[${typedSearchTerms.map((t) => t.termRep).join(', ')}]`
+}
 
 /**
 * Take a simple string as search input and process it to turn into an array of strings ready to validate and type.
@@ -219,7 +227,7 @@ export function validateAndTypeSearchTerms(searchArg: string, allowEmptyOrOnlyNe
 
   // Don't allow 0 terms, unless allowEmptyOrOnlyNegative set
   if (normalisedTerms.length === 0 && !allowEmptyOrOnlyNegative) {
-    logError(pluginJson, `No search terms submitted. Stopping.`)
+    logError('extendedSearch/validateAndTypeSearchTerms', `No search terms submitted. Stopping.`)
     return []
   }
 
@@ -251,7 +259,7 @@ export function validateAndTypeSearchTerms(searchArg: string, allowEmptyOrOnlyNe
 
   // Stop if we have a silly number of search terms
   if (validatedTerms.length > 9) {
-    logWarn(pluginJson, `Too many search terms given (${validatedTerms.length}); stopping as this might be an error.`)
+    logWarn('extendedSearch/validateAndTypeSearchTerms', `Too many search terms given (${validatedTerms.length}); stopping as this might be an error.`)
     return []
   }
 
@@ -263,7 +271,7 @@ export function validateAndTypeSearchTerms(searchArg: string, allowEmptyOrOnlyNe
       logDebug('validateAndTypeSearchTerms', 'No positive match search terms given, so adding an empty one under the hood.')
       validatedTerms.push({ term: '', type: 'must', termRep: '<empty>' })
     } else {
-      logWarn(pluginJson, 'No positive match search terms given; stopping.')
+      logWarn('extendedSearch/validateAndTypeSearchTerms', 'No positive match search terms given; stopping.')
       return []
     }
   }
@@ -313,72 +321,6 @@ export function optimiseOrderOfSearchTerms(inputTerms: Array<typedSearchTerm>): 
 }
 
 /**
- * Compute difference of two arrays, by a given property value
- * from https://stackoverflow.com/a/63745126/3238281
- * translated into Flow syntax with Generics by @nmn:
- * - PropertyName is no longer just a string type. It's now a Generic type itself called P. But we constrain P such that it must be string. How is this different from just a string? Instead of being any string, P can be a specific string literal. eg. id
- * - T is also constrained. T can no longer be any arbitrary type. It must be an object type that contains a key of the type P that we just defined. It may still have other keys indicated by the ...
- * @param {<Array<T>} arr The initial array
- * @param {<Array<T>} exclude The array to remove
- * @param {string} propertyName the key of the object to match on
- * @return {Array<T>}
- * @tests in jest file
- */
-export function differenceByPropVal<P: string, T: { +[P]: mixed, ... }> (
-  arr: $ReadOnlyArray < T >,
-    exclude: $ReadOnlyArray < T >,
-      propertyName: P
-): Array < T > {
-  return arr.filter(
-    (a: T) => !exclude.find((b: T) => b[propertyName] === a[propertyName])
-  )
-}
-
-/**
- * Simple Object equality test, working for ONE-LEVEL only objects.
- * from https://stackoverflow.com/a/5859028/3238281
- * @param {Object} o1
- * @param {Object} o2
- * @returns {boolean} does o1 = o2?
- * @test in jest file
- */
-export function compareObjects(o1: Object, o2: Object): boolean {
-  for (const p in o1) {
-    if (o1.hasOwnProperty(p)) {
-      if (o1[p] !== o2[p]) {
-        return false
-      }
-    }
-  }
-  for (const p in o2) {
-    if (o2.hasOwnProperty(p)) {
-      if (o1[p] !== o2[p]) {
-        return false
-      }
-    }
-  }
-  return true
-}
-
-/**
- * Remove the 'exclude' array terms from given 'arr' array.
- * Assumes both arrays are of the same Object type, and that we will only remove
- * when all properties are equal.
- * @param {Array<Object>} arr - array to remove from
- * @param {Array<Object>} exclucde - array to remove
- * @returns {Array<Object>} arr minus exclude
- * @tests in jest file
- */
-export function differenceByObjectEquality<P: string, T: { +[P]: mixed, ... }> (
-  arr: $ReadOnlyArray < T >,
-    exclude: $ReadOnlyArray < T >
-): Array < T > {
-  return arr.filter(
-    (a: T) => !exclude.find((b: T) => compareObjects(b, a))
-  )
-}
-
-/**
  * Returns array of intersection of arrA + arrB (only for noteAndLine types)
  * (Note: could make a more generic version of this, possibly using help from
  * https://stackoverflow.com/questions/1885557/simplest-code-for-array-intersection-in-javascript)
@@ -401,12 +343,15 @@ export function noteAndLineIntersection(arrA: Array<noteAndLine>, arrB: Array<no
 }
 
 /**
- * Get string representation of multiple search terms, complete with surrounding sqaure brackets (following Google's style)
- * @param {typedSearchTerm[]} searchTerms
- * @returns {string}
+ * Count unique filenames present in array
+ * @param {Array<noteAndLine>} inArray
+ * @returns {number} of unique filenames present
+ * @test in jest file
  */
-export function getSearchTermsRep(typedSearchTerms: Array<typedSearchTerm>): string {
-  return `[${typedSearchTerms.map((t) => t.termRep).join(', ')}]`
+export function numberOfUniqueFilenames(inArray: Array<noteAndLine>): number {
+  const uniquedFilenames = inArray.map(m => m.noteFilename).filter((val, ind, arr) => arr.indexOf(val) === ind)
+  // logDebug(`- uniqued filenames: ${uniquedFilenames.length}`)
+  return uniquedFilenames.length
 }
 
 /**
@@ -423,25 +368,25 @@ export function getSearchTermsRep(typedSearchTerms: Array<typedSearchTerm>): str
  * - then apply result limit
  * 
  * TODO: ? better document the logic with negative-only searches starting with an empty 'must' term.
- * TODO: Q: Why does data filtering happen here and not in runSearchV2?
+ * TODO: Q: Why does data filtering happen here and not in runExtendedSearch?
  * 
- * Called by runSearchesV2
- * @param {Array<resultObjectTypeV3>}
+ * Called by runExtendedSearches
+ * @param {Array<resultObjectType>}
  * @param {number} resultLimit (optional; defaults to 500)
  * @param {string?} fromDateStr optional start date limit
  * @param {string?} toDateStr optional end date limit
- * @returns {resultOutputTypeV3}
+ * @returns {resultOutputType}
  * @tests in jest file
  */
 export function applySearchOperators(
-  termsResults: Array<resultObjectTypeV3>,
+  termsResults: Array<resultObjectType>,
   resultLimit: number = 500,
   fromDateStr?: string,
   toDateStr?: string,
-): resultOutputTypeV3 {
-  const mustResultObjects: Array<resultObjectTypeV3> = termsResults.filter((t) => t.searchTerm.type === 'must')
-  const mayResultObjects: Array<resultObjectTypeV3> = termsResults.filter((t) => t.searchTerm.type === 'may')
-  const notResultObjects: Array<resultObjectTypeV3> = termsResults.filter((t) => t.searchTerm.type.startsWith('not'))
+): resultOutputType {
+  const mustResultObjects: Array<resultObjectType> = termsResults.filter((t) => t.searchTerm.type === 'must')
+  const mayResultObjects: Array<resultObjectType> = termsResults.filter((t) => t.searchTerm.type === 'may')
+  const notResultObjects: Array<resultObjectType> = termsResults.filter((t) => t.searchTerm.type.startsWith('not'))
   logDebug('applySearchOperators', `Starting with ${getSearchTermsRep(termsResults.map(m => m.searchTerm))}: ${mustResultObjects.length} must terms; ${mayResultObjects.length} may terms; ${notResultObjects.length} not terms. Limiting to ${resultLimit} results. ${(fromDateStr && toDateStr) ? '- with dates from ' + fromDateStr + ' to ' + toDateStr : 'with no dates'}`)
 
   let consolidatedNALs: Array<noteAndLine> = []
@@ -467,7 +412,7 @@ export function applySearchOperators(
     // If no results by now, there's no point finding anything further, so just form up an almost-empty return
     if (consolidatedLineCount === 0) {
       logInfo('applySearchOperators', `- must: no results found after must term [${r.searchTerm.termRep}] so stopping early.`)
-      const consolidatedResultsObject: resultOutputTypeV3 = {
+      const consolidatedResultsObject: resultOutputType = {
         searchTermsRepArr: termsResults.map((m) => m.searchTerm.termRep),
         resultNoteAndLineArr: [],
         resultCount: 0,
@@ -505,7 +450,7 @@ export function applySearchOperators(
       // If no results by now, there's no point finding anything further, so just form up an almost-empty return
       if (consolidatedLineCount === 0) {
         logInfo('applySearchOperators', `- must: no results found after must term [${r.searchTerm.termRep}] so stopping early.`)
-        const consolidatedResultsObject: resultOutputTypeV3 = {
+        const consolidatedResultsObject: resultOutputType = {
           searchTermsRepArr: termsResults.map((m) => m.searchTerm.termRep),
           resultNoteAndLineArr: [],
           resultCount: 0,
@@ -626,7 +571,7 @@ export function applySearchOperators(
   }
 
   // Form the output data structure
-  const consolidatedResultsObject: resultOutputTypeV3 = {
+  const consolidatedResultsObject: resultOutputType = {
     searchTermsRepArr: termsResults.map((m) => m.searchTerm.termRep),
     resultNoteAndLineArr: consolidatedNALs,
     resultCount: consolidatedLineCount,
@@ -660,19 +605,7 @@ export function reduceNoteAndLineArray(inArray: Array<noteAndLine>): Array<noteA
 }
 
 /**
- * Count unique filenames present in array
- * @param {Array<noteAndLine>} inArray
- * @returns {number} of unique filenames present
- * @test in jest file
- */
-export function numberOfUniqueFilenames(inArray: Array<noteAndLine>): number {
-  const uniquedFilenames = inArray.map(m => m.noteFilename).filter((val, ind, arr) => arr.indexOf(val) === ind)
-  // logDebug(`- uniqued filenames: ${uniquedFilenames.length}`)
-  return uniquedFilenames.length
-}
-
-/**
- * Run a search over all search terms in 'termsToMatchArr' over the set of notes determined by the parameters.
+ * Run an extended search over all search terms in 'termsToMatchArr' over the set of notes determined by the parameters.
  * V3 of this function, which assumes the order of terms in termsToMatchArr has been optimised.
  * Has an optional 'paraTypesToInclude' parameter of paragraph type(s) to include (e.g. ['open'] to include only open tasks). If not given, then no paragraph types will be excluded.
  *
@@ -686,9 +619,9 @@ export function numberOfUniqueFilenames(inArray: Array<noteAndLine>): number {
  * @param {boolean?} fullWord? searching (default is same as NP which is false)
  * @param {string?} fromDateStr optional start date limit to pass to applySearchOperators
  * @param {string?} toDateStr optional end date limit to pass to applySearchOperators
- * @returns {resultOutputTypeV3} results optimised for output
+ * @returns {resultOutputType} results optimised for output
  */
-export async function runSearchesV2(
+export async function runExtendedSearches(
   termsToMatchArr: Array<typedSearchTerm>,
   noteTypesToInclude: Array<string>,
   foldersToInclude: Array<string>,
@@ -697,13 +630,13 @@ export async function runSearchesV2(
   paraTypesToInclude?: Array<ParagraphType> = [],
   fromDateStr?: string,
   toDateStr?: string,
-): Promise<resultOutputTypeV3> {
+): Promise<resultOutputType> {
   try {
-    const termsResults: Array<resultObjectTypeV3> = []
+    const termsResults: Array<resultObjectType> = []
     let resultCount = 0
     let outerStartTime = new Date()
-    logDebug('runSearchesV2', `Starting with ${termsToMatchArr.length} search term(s) and paraTypes '${String(paraTypesToInclude)}'. ${config.caseSensitiveSearching ? 'caseSensitive ON. ' : ''}${config.fullWordSearching ? 'fullWord ON. ' : ''}(With ${(fromDateStr && toDateStr) ? fromDateStr + '-' + toDateStr : 'no'} dates.)`)
-    logDebug('runSearchesV2', `- config.syncOpenResultItems: ${String(config.syncOpenResultItems)}`)
+    logDebug('runExtendedSearches', `Starting with ${termsToMatchArr.length} search term(s) and paraTypes '${String(paraTypesToInclude)}'. ${config.caseSensitiveSearching ? 'caseSensitive ON. ' : ''}${config.fullWordSearching ? 'fullWord ON. ' : ''}(With ${(fromDateStr && toDateStr) ? fromDateStr + '-' + toDateStr : 'no'} dates.)`)
+    logDebug('runExtendedSearches', `- config.syncOpenResultItems: ${String(config.syncOpenResultItems)}`)
     // Now optimise the order we tackle the search terms
     const orderedSearchTerms = optimiseOrderOfSearchTerms(termsToMatchArr)
 
@@ -712,32 +645,32 @@ export async function runSearchesV2(
     // let lastTermType = ''
     for (const typedSearchTerm of orderedSearchTerms) {
       const thisTermType = typedSearchTerm.type
-      logDebug('runSearchesV2', `  - searching for term [${typedSearchTerm.termRep}] type '${thisTermType}'`)
+      logDebug('runExtendedSearches', `  - searching for term [${typedSearchTerm.termRep}] type '${thisTermType}'`)
       const innerStartTime = new Date()
 
       // do search for this search term, using configured options
-      const resultObject: resultObjectTypeV3 =
-        await runSearchV2(typedSearchTerm, noteTypesToInclude, foldersToInclude, foldersToExclude, config, paraTypesToInclude)
+      const resultObject: resultObjectType =
+        await runExtendedSearch(typedSearchTerm, noteTypesToInclude, foldersToInclude, foldersToExclude, config, paraTypesToInclude)
 
       // Save this search term and results as a new object in results array
       termsResults.push(resultObject)
       resultCount += resultObject.resultCount
-      logTimer('runSearchesV2', innerStartTime, `  -> ${resultObject.resultCount} results for '${typedSearchTerm.termRep}'`)
+      logTimer('runExtendedSearches', innerStartTime, `  -> ${resultObject.resultCount} results for '${typedSearchTerm.termRep}'`)
 
       // If we have no results from previous 'must' term, then return early
       if (thisTermType === 'must' && resultCount === 0) {
-        logInfo('runSearchesV2', `- no results from 'must' term [${typedSearchTerm.termRep}], so not doing further searches.`)
+        logInfo('runExtendedSearches', `- no results from 'must' term [${typedSearchTerm.termRep}], so not doing further searches.`)
         break
       }
     }
 
-    logTimer('runSearchesV2', outerStartTime, `- ${orderedSearchTerms.length} searches completed -> ${resultCount} results`)
+    logTimer('runExtendedSearches', outerStartTime, `- ${orderedSearchTerms.length} searches completed -> ${resultCount} results`)
 
     //------------------------------------------------------------------
     // Work out what subset of results to return, taking into the must/may/not terms, and potentially dates too
     outerStartTime = new Date()
-    const consolidatedResultSet: resultOutputTypeV3 = applySearchOperators(termsResults, config.resultLimit, fromDateStr, toDateStr)
-    logTimer('runSearchesV2', outerStartTime, `- Applied search logic`)
+    const consolidatedResultSet: resultOutputType = applySearchOperators(termsResults, config.resultLimit, fromDateStr, toDateStr)
+    logTimer('runExtendedSearches', outerStartTime, `- Applied search logic`)
 
     // For open tasks, add line sync with blockIDs (if wanted, and using NotePlan display style)
     // clo(consolidatedResultSet, 'after applySearchOperators, consolidatedResultSet =')
@@ -750,14 +683,14 @@ export async function runSearchesV2(
     }
   }
   catch (err) {
-    logError('runSearchesV2', err.message)
+    logError('runExtendedSearches', err.message)
     return { searchTermsRepArr: [], resultNoteAndLineArr: [], resultCount: 0, resultNoteCount: 0, fullResultCount: 0 } // for completeness
   }
 }
 
 /**
  * Run a search for 'searchTerm' over the set of notes determined by the parameters.
- * Returns a special resultObjectTypeV3 data structure: {
+ * Returns a special resultObjectType data structure: {
  *   searchTerm: typedSearchTerm
  *   resultNoteAndLineArr: Array<noteAndLine>  -- note: array
  *   resultCount: number
@@ -771,16 +704,16 @@ export async function runSearchesV2(
  * @param {Array<string>} foldersToExclude (can be empty list)
  * @param {SearchConfig} config object for various settings
  * @param {Array<ParagraphType>?} paraTypesToInclude optional list of paragraph types to include (e.g. 'open'). If not given, then no paragraph types will be excluded.
- * @returns {resultOutputTypeV3} combined result set optimised for output
+ * @returns {resultOutputType} combined result set optimised for output
  */
-export async function runSearchV2(
+export async function runExtendedSearch(
   typedSearchTerm: typedSearchTerm,
   noteTypesToInclude: Array<string>,
   foldersToInclude: Array<string>,
   foldersToExclude: Array<string>,
   config: SearchConfig,
   paraTypesToInclude?: Array<ParagraphType> = [],
-): Promise<resultObjectTypeV3> {
+): Promise<resultObjectType> {
   try {
     // const headingMarker = '#'.repeat(config.headingLevel)
     const fullSearchTerm = typedSearchTerm.term
@@ -791,7 +724,7 @@ export async function runSearchV2(
     const caseSensitive: boolean = config.caseSensitiveSearching
     const fullWord: boolean = config.fullWordSearching
 
-    logDebug('runSearchV2', `Starting for [${searchTerm}] with caseSensitive ${String(caseSensitive)}`)
+    logDebug('runExtendedSearch', `Starting for [${searchTerm}] with caseSensitive ${String(caseSensitive)}`)
 
     // V1: get list of matching paragraphs for this string by n.paragraphs.filter
     // ...
@@ -806,14 +739,14 @@ export async function runSearchV2(
       // use the longest word not just the first
       const longestWord = words.length > 0 ? words.sort((a, b) => b.length - a.length)[0] : ''
       searchTerm = longestWord
-      logDebug('runSearchV2', `multi-word: will just use [${searchTerm}] for [${fullSearchTerm}], and then do fuller check on results`)
+      logDebug('runExtendedSearch', `multi-word: will just use [${searchTerm}] for [${fullSearchTerm}], and then do fuller check on results`)
     }
 
     // if search term includes * or ? then we need to do further wildcard filtering: for now reduce search term to just the part before the wildcard. We will do more filtering later.
     if (searchTerm.includes("*") || searchTerm.includes("?")) {
       searchTerm = searchTerm.split(/[\*\?]/, 1)[0]
       wildcardedSearch = true
-      logDebug('runSearchV2', `wildcard: will now use [${searchTerm}] for [${fullSearchTerm}]`)
+      logDebug('runExtendedSearch', `wildcard: will now use [${searchTerm}] for [${fullSearchTerm}]`)
     }
 
     //-------------------------------------------------------
@@ -829,16 +762,16 @@ export async function runSearchV2(
     // If we have a multi-word search, then filter out the results to those that just contain the full search term
     // Same filter applies if we want case-sensitive searching
     if (multiWordSearch || caseSensitive) {
-      logDebug('runSearchV2', `multi-word or case-sensitive: before filtering: ${String(tempResult.length)}`)
+      logDebug('runExtendedSearch', `multi-word or case-sensitive: before filtering: ${String(tempResult.length)}`)
       tempResult = tempResult.filter(tr => tr.content.includes(fullSearchTerm))
-      logDebug('runSearchV2', `multi-word or case-sensitive: after filtering: ${String(tempResult.length)}`)
+      logDebug('runExtendedSearch', `multi-word or case-sensitive: after filtering: ${String(tempResult.length)}`)
     }
 
     // If we want only full word matches, then filter out the results to just those
     if (fullWord) {
-      logDebug('runSearchV2', `fullWord: before filtering: ${String(tempResult.length)}`)
+      logDebug('runExtendedSearch', `fullWord: before filtering: ${String(tempResult.length)}`)
       tempResult = tempResult.filter(tr => fullWordMatch(fullSearchTerm, tr.content, caseSensitive))
-      logDebug('runSearchV2', `fullWord: after filtering: ${String(tempResult.length)}`)
+      logDebug('runExtendedSearch', `fullWord: after filtering: ${String(tempResult.length)}`)
     }
 
     // If search term includes * or ? then we need to do further wildcard filtering, using regex equivalent:
@@ -846,9 +779,9 @@ export async function runSearchV2(
     // - replace * with [^\s]*? (i.e. any anything within the same 'word')
     if (wildcardedSearch) {
       const regexSearchTerm = new RegExp('\\b' + fullSearchTerm.replace(/\?/g, '.').replace(/\*/g, '[^\\s]*?') + '\\b')
-      logDebug('runSearchV2', `wildcard: before regex filtering with ${String(regexSearchTerm)}: ${String(tempResult.length)}`)
+      logDebug('runExtendedSearch', `wildcard: before regex filtering with ${String(regexSearchTerm)}: ${String(tempResult.length)}`)
       tempResult = tempResult.filter(tr => regexSearchTerm.test(tr.content))
-      logDebug('runSearchV2', `wildcard: after filtering: ${String(tempResult.length)}`)
+      logDebug('runExtendedSearch', `wildcard: after filtering: ${String(tempResult.length)}`)
     }
 
     if (paraTypesToInclude.length > 0) {
@@ -860,7 +793,7 @@ export async function runSearchV2(
         }
       }
       CommandBar.showLoading(false)
-      logDebug('runSearchV2', `- found ${resultParas.length} open tasks to work from`)
+      logDebug('runExtendedSearch', `- found ${resultParas.length} open tasks to work from`)
     } else {
       resultParas = tempResult
     }
@@ -868,7 +801,7 @@ export async function runSearchV2(
     const noteAndLineArr: Array<noteAndLine> = []
 
     if (resultParas.length > 0) {
-      logDebug('runSearchV2', `- Found ${resultParas.length} results for '${searchTerm}'`)
+      logDebug('runExtendedSearch', `- Found ${resultParas.length} results for '${searchTerm}'`)
 
       // Try creating much smaller data sets, without full Note or Para. Use filename for disambig later.
       let resultFieldSets: Array<reducedFieldSet> = resultParas.map((p) => {
@@ -890,13 +823,13 @@ export async function runSearchV2(
 
       // Drop out search results with the wrong paragraph type (if any given)
       let filteredParas: Array<reducedFieldSet> = []
-      logDebug('runSearchV2', `- before types filter (${paraTypesToInclude.length} = '${String(paraTypesToInclude)}')`)
+      logDebug('runExtendedSearch', `- before types filter (${paraTypesToInclude.length} = '${String(paraTypesToInclude)}')`)
       if (paraTypesToInclude && paraTypesToInclude.length > 0) {
         filteredParas = resultFieldSets.filter((p) => paraTypesToInclude.includes(p.type))
-        logDebug('runSearchV2', `- after types filter (to ${String(paraTypesToInclude)}), ${filteredParas.length} results`)
+        logDebug('runExtendedSearch', `- after types filter (to ${String(paraTypesToInclude)}), ${filteredParas.length} results`)
       } else {
         filteredParas = resultFieldSets
-        logDebug('runSearchV2', `- no type filtering requested`)
+        logDebug('runExtendedSearch', `- no type filtering requested`)
       }
 
       // Drop out search results found only in a URL or the path of a [!][link](path)
@@ -904,13 +837,13 @@ export async function runSearchV2(
         .filter((f) => !isTermInURL(searchTerm, f.content))
         .filter((f) => !isTermInMarkdownPath(searchTerm, f.content))
       if (resultFieldSets.length !== filteredParas.length) {
-        logDebug('runSearchV2', `  - URL/path filtering removed ${String(filteredParas.length - resultFieldSets.length)} results`)
+        logDebug('runExtendedSearch', `  - URL/path filtering removed ${String(filteredParas.length - resultFieldSets.length)} results`)
       }
 
       // Dedupe identical synced lines (if wanted)
-      logDebug('runSearchV2', `- Before dedupe, ${resultParas.length} results for '${searchTerm}'`)
+      logDebug('runExtendedSearch', `- Before dedupe, ${resultParas.length} results for '${searchTerm}'`)
       resultParas = eliminateDuplicateSyncedParagraphs(resultParas, 'most-recent')
-      logDebug('runSearchV2', `- After dedupe, ${resultParas.length} results for '${searchTerm}'`)
+      logDebug('runExtendedSearch', `- After dedupe, ${resultParas.length} results for '${searchTerm}'`)
 
       // Look-up table for sort details
       const sortMap = new Map([
@@ -922,7 +855,7 @@ export async function runSearchV2(
         ['created (oldest note first)', ['createdDate', 'lineIndex']],
       ])
       const sortKeys = sortMap.get(config.sortOrder) ?? 'title' // get value, falling back to 'title'
-      logDebug('runSearchV2', `- Will use sortKeys: [${String(sortKeys)}] from ${config.sortOrder}`)
+      logDebug('runExtendedSearch', `- Will use sortKeys: [${String(sortKeys)}] from ${config.sortOrder}`)
       const sortedFieldSets: Array<reducedFieldSet> = sortListBy(resultFieldSets, sortKeys)
 
       // Form the return object from sortedFieldSets
@@ -935,9 +868,9 @@ export async function runSearchV2(
       }
     }
     const resultCount = noteAndLineArr.length
-    logDebug('runSearchV2', `- end of runSearchV2 for [${searchTerm}]: ${resultCount} results from ${numberOfUniqueFilenames(noteAndLineArr)} notes`)
+    logDebug('runExtendedSearch', `- end of runExtendedSearch for [${searchTerm}]: ${resultCount} results from ${numberOfUniqueFilenames(noteAndLineArr)} notes`)
 
-    const returnObject: resultObjectTypeV3 = {
+    const returnObject: resultObjectType = {
       searchTerm: typedSearchTerm,
       resultNoteAndLineArr: noteAndLineArr,
       resultCount: resultCount,
@@ -945,7 +878,7 @@ export async function runSearchV2(
     return returnObject
   }
   catch (err) {
-    logError('runSearchV2', err.message)
+    logError('runExtendedSearch', err.message)
     // const emptyResultObject = { searchTerm: '', resultsLines: [], resultCount: 0 }
     // $FlowFixMe[incompatible-return]
     return null // for completeness
@@ -955,10 +888,10 @@ export async function runSearchV2(
 /**
  * Create a string to display the number of results and notes: "[first N] from M results from P notes"
  * @author @jgclark
- * @param {resultOutputTypeV3} resultSet
+ * @param {resultOutputType} resultSet
  * @returns {string}
  */
-export function resultCounts(resultSet: resultOutputTypeV3): string {
+export function resultCounts(resultSet: resultOutputType): string {
   return (resultSet.resultCount < resultSet.fullResultCount)
     ? `(first ${resultSet.resultCount} from ${resultSet.fullResultCount} results from ${resultSet.resultNoteCount} notes)`
     : `(${resultSet.resultCount} results from ${resultSet.resultNoteCount} notes)`
@@ -970,7 +903,7 @@ export function resultCounts(resultSet: resultOutputTypeV3): string {
  * Note: It's now possible to give a 'heading' parameter: if it's given then just that section will be replaced, otherwise the whole contents will be deleted first.
  * @author @jgclark
  *
- * @param {resultOutputTypeV3} resultSet object
+ * @param {resultOutputType} resultSet object
  * @param {string} requestedTitle requested note title to use/make
  * @param {string} titleToMatch partial title to match against existing note titles
  * @param {SearchConfig} config
@@ -979,7 +912,7 @@ export function resultCounts(resultSet: resultOutputTypeV3): string {
  * @returns {string} filename of note we've written to
  */
 export async function writeSearchResultsToNote(
-  resultSet: resultOutputTypeV3,
+  resultSet: resultOutputType,
   requestedTitle: string,
   titleToMatch: string,
   config: SearchConfig,
@@ -1059,7 +992,7 @@ export async function writeSearchResultsToNote(
  * @param {SearchConfig} config
  * @returns {Array<string>} formatted search reuslts
  */
-export function createFormattedResultLines(resultSet: resultOutputTypeV3, config: SearchConfig): Array<string> {
+export function createFormattedResultLines(resultSet: resultOutputType, config: SearchConfig): Array<string> {
   try {
     const resultOutputLines: Array<string> = []
     const headingMarker = '#'.repeat(config.headingLevel + 1)
@@ -1114,7 +1047,7 @@ export function createFormattedResultLines(resultSet: resultOutputTypeV3, config
  * @param {resultOutputTypeV2} resultSet
  * @param {SearchConfig} config
  */
-export function logBasicResultLines(resultSet: resultOutputTypeV3, config: SearchConfig): void {
+export function logBasicResultLines(resultSet: resultOutputType, config: SearchConfig): void {
   try {
     const resultOutputLines: Array<string> = []
     const simplifyLine = true
@@ -1145,10 +1078,10 @@ export function logBasicResultLines(resultSet: resultOutputTypeV3, config: Searc
 /**
  * Go through results, and if there are open task lines, then sync lines by adding a blockID (having checked there isn't one already).
  * @author @jgclark
- * @param {resultOutputTypeV3} input
- * @returns {resultOutputTypeV3}
+ * @param {resultOutputType} input
+ * @returns {resultOutputType}
  */
-export async function makeAnySyncs(input: resultOutputTypeV3): Promise<resultOutputTypeV3> {
+export async function makeAnySyncs(input: resultOutputType): Promise<resultOutputType> {
   try {
     // Go through each line looking for open tasks
     const linesToSync = []